scripts/requirements.txt
google-api-python-client==2.72.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
